feat: Clear all search history in SearchScreen
This commit updates the `SearchScreen` to add a new action that allows users to clear all search history. The action is triggered by tapping on the "Clear all history" button, which calls the `clearAllSearchHistory` method in the `SearchCubit`. This feature provides users with the ability to easily remove all search history, improving the overall user experience.