feat: Add support for AppointmentModel and PatientModel in HiveManager
This commit adds support for the `AppointmentModel` and `PatientModel` classes in the `HiveManager` class. These models are registered as adapters in the `HiveManager` to enable caching and retrieval of appointment and patient data. This enhancement improves the functionality of the application by allowing efficient storage and retrieval of appointment and patient data using Hive.