feat: Update dependencies in pubspec.yaml
This commit updates the dependencies in the pubspec.yaml file. The intl dependency is updated from version 0.18.1 to 0.19.0, and the fading_edge_scrollview dependency is overridden to version 4.1.1. These updates enhance the functionality and compatibility of the application by incorporating the latest features and bug fixes from the updated dependencies.