refactor: Add width and height properties to CustomTextButton
This commit adds the `width` and `height` properties to the `CustomTextButton` widget. These properties allow the developer to specify the desired width and height of the button, providing more flexibility in customizing the button's size. This enhancement improves the overall user experience by allowing better control over the layout and design of the button.